+++ include/nstd/shared_lib.h
@@ -4,17 +4,17 @@
 #include "core/str.h"
 #include "nstd.h"
 #include "os/os.h"
-#ifdef NSTD_OS_WINDOWS
+#if NSTD_OS_WINDOWS
 #    include "os/windows/shared_lib.h"
 #endif
 
 /// A handle to a dynamically loaded library.
-#if defined(NSTD_OS_UNIX)
+#if NSTD_OS_UNIX
 typedef struct {
     /// A raw handle to the shared library.
     NSTDAnyMut handle;
 } NSTDSharedLib;
-#elif defined(NSTD_OS_WINDOWS)
+#elif NSTD_OS_WINDOWS
 typedef NSTDWindowsSharedLib NSTDSharedLib;
 #else
 typedef NSTDAnyMut NSTDSharedLib;